feature: split class and function solvers Handles https://github.com/PyLops/pylops/issues/390.
